Fix: Change tg_id from Integer to BigInteger to support large Telegram user IDs

+++ models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, ForeignKey, DateTime, Boolean, Time, Enum as SQLEnum
+from sqlalchemy import Column, Integer, BigInteger, String, ForeignKey, DateTime, Boolean, Time, Enum as SQLEnum
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 import enum
@@ -24,7 +24,7 @@ class User(Base):
     __tablename__ = "users"
 
     id = Column(Integer, primary_key=True, index=True)
-    tg_id = Column(Integer, unique=True, index=True, nullable=False)
+    tg_id = Column(BigInteger, unique=True, index=True, nullable=False)
     role = Column(SQLEnum(UserRole), nullable=False)
     timezone = Column(String, default="UTC")
